Revert "M320-D2 TEV changes"
This reverts commit 49706f19

+++ M320-D2/src/Test.java
@@ -26,6 +26,6 @@ public class Test {
     }
 
     public float calculateGrade() {
-        return (((float) points / maxPoints * 5) + 1);
+        return ((points / maxPoints * 5) + 1);
     }
 }
